🧪 GA1SAK tests: split PR open flow suite
