+++ .agentplane/tasks/202605010645-GA1SAK/pr/github-body.md
@@ -1,0 +1,33 @@
+## Summary
+
+AP-12: Split PR open flow tests
+
+Split PR open flow tests by artifact, git, validation, and network gates using cli-core-pr-flow testkit helpers.
+
+## Scope
+
+- In scope: Split PR open flow tests by artifact, git, validation, and network gates using cli-core-pr-flow testkit helpers.
+- Out of scope: unrelated refactors not required for "AP-12: Split PR open flow tests".
+
+## Verification
+
+- State: ok
+- Note: Focused PR-open split verification passed: PR-flow suites, routing/inventory checks, oversized baseline, typecheck, lint, formatting, bootstrap, doctor, and policy routing were green.
+- Full verification checklist lives in local review.md.
+
+## Handoff Notes
+
+- No handoff notes recorded yet. Use `agentplane pr note ...` to append one.
+
+<details>
+<summary>Raw evidence</summary>
+
+- Updated: 2026-05-01T11:02:36.968Z
+- Branch: task/202605010645-GA1SAK/pr-open-flow-test-split
+- Head: c389f1be9734
+
+```text
+No changes detected.
+```
+
+</details>
