+++ packages/agentplane/src/cli/run-cli.core.pr-flow.pr-open.artifacts.test.ts
@@ -1,0 +1,213 @@
+/* eslint-disable @typescript-eslint/no-unused-vars */
+import { describe } from "vitest";
+
+import {
+  PR_FLOW_INTEGRATION_TIMEOUT_MS,
+  PR_FLOW_LONG_TIMEOUT_MS,
+  approveTaskPlan,
+  branchPrArtifactFixture,
+  captureStdIO,
+  chmod,
+  cleanGitEnv,
+  commitAll,
+  commitPathsIfChanged,
+  configureGitUser,
+  configurePushableOrigin,
+  createUpgradeBundle,
+  defaultConfig,
+  execFile,
+  expect,
+  extractTaskSuffix,
+  filterAgentsByWorkflow,
+  getAgentplaneHome,
+  gitBranchExists,
+  it,
+  loadAgentTemplates,
+  loadAgentsTemplate,
+  mkdir,
+  mkGitRepoRoot,
+  mkGitRepoRootWithBranch,
+  mkTempDir,
+  mkdtemp,
+  os,
+  path,
+  pathExists,
+  promisify,
+  prompts,
+  readFile,
+  readFileSync,
+  readTask,
+  realpath,
+  readdir,
+  recordVerificationOk,
+  renderTaskReadme,
+  resolveUpdateCheckCachePath,
+  rm,
+  runCli,
+  runCliSilent,
+  setConcreteVerifySteps,
+  stageGitignoreIfPresent,
+  stubTaskBackend,
+  validateCommitSubject,
+  vi,
+  writeConfig,
+  writeDefaultConfig,
+  writeFile,
+  installFakeGhPrApi,
+  installFakeGhPrApiRequiringPublishedPacketHead,
+  installFakeGhPrLookup,
+  type ResolvedProject,
+} from "@agentplane/testkit/cli-core-pr-flow";
+
+describe("runCli pr open flow artifacts", { timeout: PR_FLOW_INTEGRATION_TIMEOUT_MS }, () => {
+  it("pr open creates PR artifacts", async () => {
+    const root = await mkGitRepoRootWithBranch("main");
+    const config = defaultConfig();
+    config.workflow_mode = "branch_pr";
+    await writeConfig(root, config);
+
+    let taskId = "";
+    const ioTask = captureStdIO();
+    try {
+      const code = await runCli([
+        "task",
+        "new",
+        "--title",
+        "PR open task",
+        "--description",
+        "PR open creates artifacts",
+        "--priority",
+        "med",
+        "--owner",
+        "CODER",
+        "--tag",
+        "nodejs",
+        "--root",
+        root,
+      ]);
+      expect(code).toBe(0);
+      taskId = ioTask.stdout.trim();
+    } finally {
+      ioTask.restore();
+    }
+
+    const io = captureStdIO();
+    try {
+      const code = await runCli([
+        "pr",
+        "open",
+        taskId,
+        "--author",
+        "CODER",
+        "--branch",
+        `task/${taskId}/pr-open`,
+        "--root",
+        root,
+      ]);
+      expect(code).toBe(0);
+      expect(io.stdout).toContain("✅ pr open");
+      expect(io.stdout).toContain(
+        "local PR artifacts synced; remote PR creation staged (GitHub origin repo unavailable)",
+      );
+    } finally {
+      io.restore();
+    }
+
+    const prArtifacts = branchPrArtifactFixture(root, taskId);
+    const meta = await prArtifacts.readMeta<{
+      task_id?: string;
+      branch?: string;
+      artifact_state?: string;
+      artifact_state_reason?: string;
+    }>();
+    expect(meta.task_id).toBe(taskId);
+    expect(meta.branch).toBe(`task/${taskId}/pr-open`);
+    expect(meta.artifact_state).toBe("remote_staged");
+    expect(meta.artifact_state_reason).toBe("GitHub origin repo unavailable");
+    expect(await prArtifacts.readReview()).toContain("## Scope");
+    await readFile(prArtifacts.diffstatPath, "utf8");
+    expect(await readFile(prArtifacts.notesPath, "utf8")).toBe("");
+    await readFile(prArtifacts.verifyLogPath, "utf8");
+    expect(await readFile(prArtifacts.githubTitlePath, "utf8")).toContain(
+      `(${extractTaskSuffix(taskId)})`,
+    );
+    expect(await prArtifacts.readGithubBody()).toContain("## Verification");
+    expect(await prArtifacts.readGithubBody()).not.toContain("## Risks");
+  });
+
+  it("pr open renders diffstat immediately when the task branch already has changes", async () => {
+    const root = await mkGitRepoRootWithBranch("main");
+    const config = defaultConfig();
+    config.workflow_mode = "branch_pr";
+    await writeConfig(root, config);
+    await configureGitUser(root);
+    const execFileAsync = promisify(execFile);
+
+    await writeFile(path.join(root, "seed.txt"), "seed", "utf8");
+    await execFileAsync("git", ["add", "seed.txt"], { cwd: root });
+    await execFileAsync("git", ["commit", "-m", "seed"], { cwd: root });
+
+    let taskId = "";
+    const ioTask = captureStdIO();
+    try {
+      const code = await runCli([
+        "task",
+        "new",
+        "--title",
+        "PR open existing branch diffstat",
+        "--description",
+        "PR open should render the current branch diffstat without waiting for pr update",
+        "--priority",
+        "med",
+        "--owner",
+        "CODER",
+        "--tag",
+        "nodejs",
+        "--root",
+        root,
+      ]);
+      expect(code).toBe(0);
+      taskId = ioTask.stdout.trim();
+    } finally {
+      ioTask.restore();
+    }
+
+    await runCliSilent(["branch", "base", "set", "main", "--root", root]);
+
+    const branch = `task/${taskId}/pr-open-existing-diffstat`;
+    await execFileAsync("git", ["checkout", "-b", branch], { cwd: root });
+    await writeFile(path.join(root, "change.txt"), "change", "utf8");
+    await execFileAsync("git", ["add", "."], { cwd: root });
+    await execFileAsync("git", ["commit", "-m", "task branch change"], { cwd: root });
+
+    const io = captureStdIO();
+    try {
+      const code = await runCli([
+        "pr",
+        "open",
+        taskId,
+        "--author",
+        "CODER",
+        "--branch",
+        branch,
+        "--root",
+        root,
+      ]);
+      expect(code).toBe(0);
+      expect(io.stdout).toContain("✅ pr open");
+    } finally {
+      io.restore();
+    }
+
+    const prArtifacts = branchPrArtifactFixture(root, taskId);
+    const diffstat = await readFile(prArtifacts.diffstatPath, "utf8");
+    const review = await prArtifacts.readReview();
+    const githubBody = await prArtifacts.readGithubBody();
+
+    expect(diffstat).toContain("change.txt");
+    expect(review).toContain("change.txt");
+    expect(githubBody).toContain("change.txt");
+    expect(review).not.toContain("No changes detected.");
+    expect(githubBody).not.toContain("No changes detected.");
+  });
+});
