+++ packages/agentplane/src/cli/run-cli.core.pr-flow.pr-open.git.test.ts
@@ -58,347 +58,7 @@ import {
   type ResolvedProject,
 } from "@agentplane/testkit/cli-core-pr-flow";
 
-describe("runCli pr open flow", { timeout: PR_FLOW_INTEGRATION_TIMEOUT_MS }, () => {
-  it("pr open creates PR artifacts", async () => {
-    const root = await mkGitRepoRootWithBranch("main");
-    const config = defaultConfig();
-    config.workflow_mode = "branch_pr";
-    await writeConfig(root, config);
-
-    let taskId = "";
-    const ioTask = captureStdIO();
-    try {
-      const code = await runCli([
-        "task",
-        "new",
-        "--title",
-        "PR open task",
-        "--description",
-        "PR open creates artifacts",
-        "--priority",
-        "med",
-        "--owner",
-        "CODER",
-        "--tag",
-        "nodejs",
-        "--root",
-        root,
-      ]);
-      expect(code).toBe(0);
-      taskId = ioTask.stdout.trim();
-    } finally {
-      ioTask.restore();
-    }
-
-    const io = captureStdIO();
-    try {
-      const code = await runCli([
-        "pr",
-        "open",
-        taskId,
-        "--author",
-        "CODER",
-        "--branch",
-        `task/${taskId}/pr-open`,
-        "--root",
-        root,
-      ]);
-      expect(code).toBe(0);
-      expect(io.stdout).toContain("✅ pr open");
-      expect(io.stdout).toContain(
-        "local PR artifacts synced; remote PR creation staged (GitHub origin repo unavailable)",
-      );
-    } finally {
-      io.restore();
-    }
-
-    const prDir = path.join(root, ".agentplane", "tasks", taskId, "pr");
-    const metaRaw = await readFile(path.join(prDir, "meta.json"), "utf8");
-    const meta = JSON.parse(metaRaw) as {
-      task_id?: string;
-      branch?: string;
-      artifact_state?: string;
-      artifact_state_reason?: string;
-    };
-    expect(meta.task_id).toBe(taskId);
-    expect(meta.branch).toBe(`task/${taskId}/pr-open`);
-    expect(meta.artifact_state).toBe("remote_staged");
-    expect(meta.artifact_state_reason).toBe("GitHub origin repo unavailable");
-    expect(await readFile(path.join(prDir, "review.md"), "utf8")).toContain("## Scope");
-    await readFile(path.join(prDir, "diffstat.txt"), "utf8");
-    expect(await readFile(path.join(prDir, "notes.jsonl"), "utf8")).toBe("");
-    await readFile(path.join(prDir, "verify.log"), "utf8");
-    expect(await readFile(path.join(prDir, "github-title.txt"), "utf8")).toContain(
-      `(${extractTaskSuffix(taskId)})`,
-    );
-    expect(await readFile(path.join(prDir, "github-body.md"), "utf8")).toContain("## Verification");
-    expect(await readFile(path.join(prDir, "github-body.md"), "utf8")).not.toContain("## Risks");
-  });
-
-  it("pr open renders diffstat immediately when the task branch already has changes", async () => {
-    const root = await mkGitRepoRootWithBranch("main");
-    const config = defaultConfig();
-    config.workflow_mode = "branch_pr";
-    await writeConfig(root, config);
-    await configureGitUser(root);
-    const execFileAsync = promisify(execFile);
-
-    await writeFile(path.join(root, "seed.txt"), "seed", "utf8");
-    await execFileAsync("git", ["add", "seed.txt"], { cwd: root });
-    await execFileAsync("git", ["commit", "-m", "seed"], { cwd: root });
-
-    let taskId = "";
-    const ioTask = captureStdIO();
-    try {
-      const code = await runCli([
-        "task",
-        "new",
-        "--title",
-        "PR open existing branch diffstat",
-        "--description",
-        "PR open should render the current branch diffstat without waiting for pr update",
-        "--priority",
-        "med",
-        "--owner",
-        "CODER",
-        "--tag",
-        "nodejs",
-        "--root",
-        root,
-      ]);
-      expect(code).toBe(0);
-      taskId = ioTask.stdout.trim();
-    } finally {
-      ioTask.restore();
-    }
-
-    await runCliSilent(["branch", "base", "set", "main", "--root", root]);
-
-    const branch = `task/${taskId}/pr-open-existing-diffstat`;
-    await execFileAsync("git", ["checkout", "-b", branch], { cwd: root });
-    await writeFile(path.join(root, "change.txt"), "change", "utf8");
-    await execFileAsync("git", ["add", "."], { cwd: root });
-    await execFileAsync("git", ["commit", "-m", "task branch change"], { cwd: root });
-
-    const io = captureStdIO();
-    try {
-      const code = await runCli([
-        "pr",
-        "open",
-        taskId,
-        "--author",
-        "CODER",
-        "--branch",
-        branch,
-        "--root",
-        root,
-      ]);
-      expect(code).toBe(0);
-      expect(io.stdout).toContain("✅ pr open");
-    } finally {
-      io.restore();
-    }
-
-    const prDir = path.join(root, ".agentplane", "tasks", taskId, "pr");
-    const diffstat = await readFile(path.join(prDir, "diffstat.txt"), "utf8");
-    const review = await readFile(path.join(prDir, "review.md"), "utf8");
-    const githubBody = await readFile(path.join(prDir, "github-body.md"), "utf8");
-
-    expect(diffstat).toContain("change.txt");
-    expect(review).toContain("change.txt");
-    expect(githubBody).toContain("change.txt");
-    expect(review).not.toContain("No changes detected.");
-    expect(githubBody).not.toContain("No changes detected.");
-  });
-
-  it("pr open creates a remote GitHub PR when origin and gh are available", async () => {
-    const root = await mkGitRepoRootWithBranch("main");
-    const config = defaultConfig();
-    config.workflow_mode = "branch_pr";
-    await writeConfig(root, config);
-    await configureGitUser(root);
-    const execFileAsync = promisify(execFile);
-    await execFileAsync("git", ["remote", "add", "origin", "https://github.com/example/repo.git"], {
-      cwd: root,
-      env: cleanGitEnv(),
-    });
-    await configurePushableOrigin(root);
-    await runCliSilent(["branch", "base", "set", "main", "--root", root]);
-    await execFileAsync("git", ["commit", "--allow-empty", "-m", "chore test setup"], {
-      cwd: root,
-      env: cleanGitEnv(),
-    });
-
-    let taskId = "";
-    const ioTask = captureStdIO();
-    try {
-      const code = await runCli([
-        "task",
-        "new",
-        "--title",
-        "PR open remote create",
-        "--description",
-        "PR open should create a remote PR when GitHub is available",
-        "--priority",
-        "med",
-        "--owner",
-        "CODER",
-        "--tag",
-        "github",
-        "--root",
-        root,
-      ]);
-      expect(code).toBe(0);
-      taskId = ioTask.stdout.trim();
-    } finally {
-      ioTask.restore();
-    }
-
-    const branch = `task/${taskId}/remote-create`;
-    const { fakeBin, logPath } = await installFakeGhPrApi({
-      scenarioName: "open-create",
-      branch,
-      createResponse: {
-        number: 654,
-        html_url: "https://github.com/example/repo/pull/654",
-        state: "open",
-        merged_at: null,
-        merge_commit_sha: null,
-        head: { sha: "remote-head-sha" },
-        base: { ref: "main" },
-      },
-    });
-    const originalPath = process.env.PATH;
-    process.env.PATH = `${fakeBin}${path.delimiter}${originalPath ?? ""}`;
-    process.env.AGENTPLANE_GH_LOG = logPath;
-
-    const io = captureStdIO();
-    try {
-      const code = await runCli([
-        "pr",
-        "open",
-        taskId,
-        "--author",
-        "CODER",
-        "--branch",
-        branch,
-        "--root",
-        root,
-      ]);
-      expect(code).toBe(0);
-      expect(io.stdout).toContain("created GitHub PR #654");
-      expect(io.stdout).toContain("https://github.com/example/repo/pull/654");
-    } finally {
-      io.restore();
-      process.env.PATH = originalPath;
-      delete process.env.AGENTPLANE_GH_LOG;
-    }
-
-    const metaPath = path.join(root, ".agentplane", "tasks", taskId, "pr", "meta.json");
-    const meta = JSON.parse(await readFile(metaPath, "utf8")) as {
-      pr_number?: number;
-      pr_url?: string;
-      status?: string;
-      head_sha?: string;
-    };
-    expect(meta.pr_number).toBeUndefined();
-    expect(meta.pr_url).toBeUndefined();
-    expect(meta.status).toBeUndefined();
-    expect(meta.head_sha).toBeUndefined();
-
-    const logText = await readFile(logPath, "utf8");
-    const log = logText
-      .trim()
-      .split("\n")
-      .filter(Boolean)
-      .map((line) => JSON.parse(line) as string[]);
-    expect(
-      log.some(
-        (args) =>
-          args[0] === "api" &&
-          args[1] === "repos/example/repo/pulls" &&
-          args.includes("-X") &&
-          args.includes("POST"),
-      ),
-    ).toBe(true);
-  });
-
-  it("pr open marks remote creation failures explicitly in pr metadata", async () => {
-    const root = await mkGitRepoRootWithBranch("main");
-    const config = defaultConfig();
-    config.workflow_mode = "branch_pr";
-    await writeConfig(root, config);
-    await configureGitUser(root);
-    const execFileAsync = promisify(execFile);
-    await execFileAsync("git", ["remote", "add", "origin", "https://github.com/example/repo.git"], {
-      cwd: root,
-      env: cleanGitEnv(),
-    });
-    await configurePushableOrigin(root);
-    await runCliSilent(["branch", "base", "set", "main", "--root", root]);
-
-    let taskId = "";
-    const ioTask = captureStdIO();
-    try {
-      const code = await runCli([
-        "task",
-        "new",
-        "--title",
-        "PR open remote creation failure",
-        "--description",
-        "PR open should persist a failed remote creation state when GitHub rejects PR creation.",
-        "--priority",
-        "med",
-        "--owner",
-        "CODER",
-        "--tag",
-        "github",
-        "--root",
-        root,
-      ]);
-      expect(code).toBe(0);
-      taskId = ioTask.stdout.trim();
-    } finally {
-      ioTask.restore();
-    }
-
-    const branch = `task/${taskId}/remote-create-failure`;
-    const { fakeBin } = await installFakeGhPrApi({
-      scenarioName: "open-create-failure",
-      branch,
-      createResponse: {},
-      createError: "HTTP 403: GitHub auth unavailable",
-    });
-    const originalPath = process.env.PATH;
-    process.env.PATH = `${fakeBin}${path.delimiter}${originalPath ?? ""}`;
-
-    const io = captureStdIO();
-    try {
-      const code = await runCli([
-        "pr",
-        "open",
-        taskId,
-        "--author",
-        "CODER",
-        "--branch",
-        branch,
-        "--root",
-        root,
-      ]);
-      expect(code).toBe(0);
-      expect(io.stdout).toContain("remote PR creation failed");
-    } finally {
-      io.restore();
-      process.env.PATH = originalPath;
-    }
-
-    const meta = JSON.parse(
-      await readFile(path.join(root, ".agentplane", "tasks", taskId, "pr", "meta.json"), "utf8"),
-    ) as { artifact_state?: string; artifact_state_reason?: string };
-    expect(meta.artifact_state).toBe("remote_failed");
-    expect(meta.artifact_state_reason).toBe("GitHub auth or permissions unavailable");
-  });
-
+describe("runCli pr open flow git publishing", { timeout: PR_FLOW_INTEGRATION_TIMEOUT_MS }, () => {
   it("pr open marks push failures explicitly before returning an error", async () => {
     const root = await mkGitRepoRootWithBranch("main");
     const config = defaultConfig();
@@ -1004,211 +664,5 @@ describe("runCli pr open flow", { timeout: PR_FLOW_INTEGRATION_TIMEOUT_MS }, () 
       io.restore();
       process.env.PATH = originalPath;
     }
-  });
-
-  it("pr open ignores dotenv-injected GitHub tokens when gh auth is otherwise available", async () => {
-    const root = await mkGitRepoRootWithBranch("main");
-    const config = defaultConfig();
-    config.workflow_mode = "branch_pr";
-    await writeConfig(root, config);
-    await configureGitUser(root);
-    const execFileAsync = promisify(execFile);
-    await execFileAsync("git", ["remote", "add", "origin", "https://github.com/example/repo.git"], {
-      cwd: root,
-      env: cleanGitEnv(),
-    });
-    await configurePushableOrigin(root);
-    await runCliSilent(["branch", "base", "set", "main", "--root", root]);
-
-    let taskId = "";
-    const ioTask = captureStdIO();
-    try {
-      const code = await runCli([
-        "task",
-        "new",
-        "--title",
-        "PR open dotenv auth drift",
-        "--description",
-        "PR open should ignore dotenv-injected GitHub auth that would override gh login",
-        "--priority",
-        "med",
-        "--owner",
-        "CODER",
-        "--tag",
-        "github",
-        "--root",
-        root,
-      ]);
-      expect(code).toBe(0);
-      taskId = ioTask.stdout.trim();
-    } finally {
-      ioTask.restore();
-    }
-
-    const branch = `task/${taskId}/dotenv-auth`;
-    const { fakeBin, logPath } = await installFakeGhPrApi({
-      scenarioName: "open-dotenv-auth",
-      branch,
-      rejectEnvKey: "GITHUB_TOKEN",
-      rejectEnvValue: "dotenv-bad-token",
-      createResponse: {
-        number: 655,
-        html_url: "https://github.com/example/repo/pull/655",
-        state: "open",
-        merged_at: null,
-        merge_commit_sha: null,
-        head: { sha: "remote-head-sha" },
-        base: { ref: "main" },
-      },
-    });
-    const originalPath = process.env.PATH;
-    const originalGithubToken = process.env.GITHUB_TOKEN;
-    const originalDotenvKeys = process.env.AGENTPLANE_DOTENV_LOADED_KEYS;
-    process.env.PATH = `${fakeBin}${path.delimiter}${originalPath ?? ""}`;
-    process.env.AGENTPLANE_GH_LOG = logPath;
-    process.env.GITHUB_TOKEN = "dotenv-bad-token";
-    process.env.AGENTPLANE_DOTENV_LOADED_KEYS = "GITHUB_TOKEN";
-
-    const io = captureStdIO();
-    try {
-      const code = await runCli([
-        "pr",
-        "open",
-        taskId,
-        "--author",
-        "CODER",
-        "--branch",
-        branch,
-        "--root",
-        root,
-      ]);
-      expect(code).toBe(0);
-      expect(io.stdout).toContain("created GitHub PR #655");
-      expect(io.stdout).not.toContain("remote PR creation staged");
-    } finally {
-      io.restore();
-      process.env.PATH = originalPath;
-      delete process.env.AGENTPLANE_GH_LOG;
-      if (originalGithubToken === undefined) delete process.env.GITHUB_TOKEN;
-      else process.env.GITHUB_TOKEN = originalGithubToken;
-      if (originalDotenvKeys === undefined) delete process.env.AGENTPLANE_DOTENV_LOADED_KEYS;
-      else process.env.AGENTPLANE_DOTENV_LOADED_KEYS = originalDotenvKeys;
-    }
-
-    const metaPath = path.join(root, ".agentplane", "tasks", taskId, "pr", "meta.json");
-    const meta = JSON.parse(await readFile(metaPath, "utf8")) as {
-      pr_number?: number;
-      pr_url?: string;
-      status?: string;
-      head_sha?: string;
-    };
-    expect(meta.pr_number).toBeUndefined();
-    expect(meta.pr_url).toBeUndefined();
-    expect(meta.status).toBeUndefined();
-    expect(meta.head_sha).toBeUndefined();
-  });
-
-  it("pr open respects --sync-only and skips remote GitHub PR creation", async () => {
-    const root = await mkGitRepoRootWithBranch("main");
-    const config = defaultConfig();
-    config.workflow_mode = "branch_pr";
-    await writeConfig(root, config);
-    await configureGitUser(root);
-    const execFileAsync = promisify(execFile);
-    await execFileAsync("git", ["remote", "add", "origin", "https://github.com/example/repo.git"], {
-      cwd: root,
-      env: cleanGitEnv(),
-    });
-    await configurePushableOrigin(root);
-    await runCliSilent(["branch", "base", "set", "main", "--root", root]);
-
-    let taskId = "";
-    const ioTask = captureStdIO();
-    try {
-      const code = await runCli([
-        "task",
-        "new",
-        "--title",
-        "PR open sync only",
-        "--description",
-        "PR open should allow local-only artifact sync without remote creation",
-        "--priority",
-        "med",
-        "--owner",
-        "CODER",
-        "--tag",
-        "github",
-        "--root",
-        root,
-      ]);
-      expect(code).toBe(0);
-      taskId = ioTask.stdout.trim();
-    } finally {
-      ioTask.restore();
-    }
-
-    const branch = `task/${taskId}/sync-only`;
-    const { fakeBin, logPath } = await installFakeGhPrApi({
-      scenarioName: "open-sync-only",
-      branch,
-      createResponse: {
-        number: 987,
-        html_url: "https://github.com/example/repo/pull/987",
-        state: "open",
-        merged_at: null,
-        merge_commit_sha: null,
-        head: { sha: "remote-head-sha" },
-        base: { ref: "main" },
-      },
-    });
-    const originalPath = process.env.PATH;
-    process.env.PATH = `${fakeBin}${path.delimiter}${originalPath ?? ""}`;
-    process.env.AGENTPLANE_GH_LOG = logPath;
-
-    const io = captureStdIO();
-    try {
-      const code = await runCli([
-        "pr",
-        "open",
-        taskId,
-        "--author",
-        "CODER",
-        "--branch",
-        branch,
-        "--sync-only",
-        "--root",
-        root,
-      ]);
-      expect(code).toBe(0);
-      expect(io.stdout).toMatch(
-        /remote PR creation (skipped \(--sync-only\)|staged \(GitHub PR creation failed\))/,
-      );
-    } finally {
-      io.restore();
-      process.env.PATH = originalPath;
-      delete process.env.AGENTPLANE_GH_LOG;
-    }
-
-    const metaPath = path.join(root, ".agentplane", "tasks", taskId, "pr", "meta.json");
-    const meta = JSON.parse(await readFile(metaPath, "utf8")) as {
-      pr_number?: number;
-    };
-    expect(meta.pr_number).toBeUndefined();
-
-    const logText = await readFile(logPath, "utf8");
-    const log = logText
-      .trim()
-      .split("\n")
-      .filter(Boolean)
-      .map((line) => JSON.parse(line) as string[]);
-    expect(
-      log.some(
-        (args) =>
-          args[0] === "api" &&
-          args[1] === "repos/example/repo/pulls" &&
-          args.includes("-X") &&
-          args.includes("POST"),
-      ),
-    ).toBe(false);
   });
 });
